requirements.txt
flask==3.0.3
flask-cors==4.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
langchain==0.3.7
langchain-google-genai==2.0.7
langchain-core==0.3.15
google-generativeai==0.8.3

# v2: Web scraping
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2

# v2: Headless browser fallback for bot-protected sites
playwright==1.44.0
# After install run: playwright install chromium
